requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
googlesearch-python
openai
pyannote.audio>=2.1.1
torch>=2.0.0
pydub>=0.25.1
numpy==1.24.3
transformers>=4.30.0
matplotlib>=3.7.0
spacy
nltk
tqdm
pandas
seaborn
wordcloud
chromadb 
sentence-transformers